requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.19
Django==1.8.2
djangorestframework==3.4.0
kombu==3.0.35
psycopg2==2.6.1
pytz==2016.6.1
redis==2.10.5
wheel==0.24.0
